Avoid holding the main thread on a long trace
Whenever the trace Processor tries to read a trace file with lots of events or deep calltstacks it will hang for a long time.

This makes PerfLab non-responsive, which is bad.

By yielding to the main thread I can handle trace files past 200MB with no hickups to the UI